Use boundingBox of element to get click point
BoundingBox returns the values in page coordinates, which is what we
want. The logic is easier to follow too.

It doesn't always work though and in some cases (with iframes that are
translated is my current thinking) we still need to perform a post
process to realign the point to be in page coords rather than relative
to the frame.