Add step into frame logic to waitForSelector
When we wait for a selector, we need to be able to step into an
iframe/frame when a directive in the selector is found. This change
performs those steps. It will split the selector, retrieve the document
for the iframe/frame and then continue the second selector split after
stepping into the iframe/frame.