Add step in iframe in count
We need to step into iframe/frame when working with count too. This
will split on the frame directive, step into the frame by splitting the
selector in two, and using the first split to do the step in. The
second split of the selector will be used after stepping into the frame