grpc: Fix usage of Any types after refactor on protobuf loading (#5089)
After #5017 turned out that k6 does not propagate the types it loads when making gRPC calls. This means that in many of the cases the globalTypes will be used, which will lead to errors.

This likely is not just around Any but is how it was reported.

This commit goes and propagates the types in each place where 
they will be marshaled/unmarshalled.

The added test is not perfect, but does catch the problem without
needing #4981 which I would expect would've also caught this.
As currently the server will load the prototypes in the global types
and k6/net/grpc will load them in their local types.
And then they will technically work.