Update ElementHandle.BoundingBox to return err
This method was designed not to return an error since if an error
occurred we assumed that the element was hidden, and so we couldn't get
a viable bounding box for the element.

I've changed it so that now it returns an error. Areas of the codebase
that should interpret errors as not an error should handle that them
selves.