Update waitForElementStates for stability state
We want to wait on stability first before continuing on with the other
state checks. An element is regarded as stable when 10 RAF frames have
passed and the element's rect is in the same place. If the element is
hidden in one call, and another then the rect will not match and
therefore we cannot say the element is stable. The function before
allowed this to occur i.e. if the element was hidden and unhidden the
counter wouldn't be reset and eventually it would mark the element as
stable even though it isn't.