Update log from warn to debug
There's nothing that the user can do with this warning, so lets change
it to debug to avoid confusion. It's an internal issue that we should
look into when users want to work with other types. At the moment we
support page and background_page. We might need to work with
service_worker.