modules: refactor to be able to load modules without running
Let us have a module resolver that just loads a main module and all of
its dependencies without running an js code.

This is useful if we need to do more checks or something else and we
need the modules that will be loaded but not actually run it.

Will be used for #5166 among other stuff including removing k6deps and
esbuild for doing the auto extension resolution.