Refactor "error"notvisible" to error type
This is just refactoring and removing the need to work with a string
error. I've double checked our codebase, playwright and Chromium to
make sure that this error doesn't originate from something outside.
The string error is something internal to the browser module, so it is
safe to replace with the concrete type.