Preload modules from `require` calls
Now when a loaded js module has a `require` function call its argument
(if string) will be loaded as a module and we will repeat the process.

This should load all modules, and more importantly will help with
finding if any missing modules can not be loaded.

This is very likely not a good idea in general, but is what esbuild is
doing currently if auto extension resolution is enabled. So this is
copying this behaviour.

It is likely that this will be disabled in the future (v2) and might
need or not need additional testing.