fix clearTimeout running next timer if first is removed
clearTimeout was not clearing the first timer if it was removing it.
Leading to it executing and pulling the next queued item. Which
potentially can run a much later queued item.

Now the clearTimeout specifically checks and if it is removing the
first timer it will also reset so it will correctly run the remaining
ones when they should be.

Fixes #5199