Refactor auto extension resolution to not use k6deps and esbuild (#5320)
Reimplement auto extension resolution without using esbuild or k6deps,
but the internal methods that k6 uses to load modules.


This removes the need for a separate project to have the same way to parse
 and understand imports and also double parsing everything.

Also fixing a number of bugs coming from this and removes the maintenance
burden to keep both projects align. 

It also fixes problems with paths on windows as it is now using the internal handling everywhere.

Makes use directives actually only work at the beginning of files, but also for all files.

It does now not work with `require` calls, at least until/if  #5319 is merged.

 Correctly propagates exit codes from sub processes up.

Fixes #5127
FIxes #5176
Fixes #5104
Closes #5102

Likely some updates to every other issue around automatic extension resolution as well.

Co-authored-by: Ivan <2103732+codebien@users.noreply.github.com>
Co-authored-by: Théo Crevon <oleiade@users.noreply.github.com>