feat(deps): make it work on Windows too
This was a bit cumbersome because of the multiple ways of
setting the work environemnt (local vs MinGW Go/make installation).

I don't think we should handle all the cases and each user that runs
`make generate` should be able to fix their environment to make it work.

I've used MinGW64 + local Go installation and it's working.