Allow releasing from vx.y.z-release[N] tags
This for releasing from tags when we make a mistake while adding a tag.
For example, if we add the version tag v1.4.1 and then forget to include
something, we still want to release for v1.4.1. To do that, we can
create additional tags like v1.4.1-release-1, v1.4.1-release-2, etc. so
that we can release off of these tags since they will be considered as
a v1.4.1 release, not a v1.4.1-release-N.