Fix build workflow publish-packages auth issue
Fixes the authentication error in the `publish-packages` job when
logging into GitHub Container Registry:

The `publish-packages` job fails with "Username and password
required" when attempting to login to ghcr.io. This happens because
the job tries to use `${{ env.GITHUB_ACTOR }}` and
`${{ env.GITHUB_TOKEN }}` environment variables that are never set.