Improve circular object logging
Previously, we were returning false to avoid traversing over circular
objects. However, we were printing [object Object] for the whole object
(or the sobek.Value's string representation). We're now returning a
Circular string when confronted with circularity. This removes the need
to keep track of returning a bool to stop traversing, as we return a
string, which is not traversable, and the algorithm will stop by itself.