chore: Fix tests and subcommand handling in `version` command
The addition of subcommand extension did not handle the case in
versions. As such if there are subcommand extensions, version fails.

In tests this might happen if the test for subcommand runs before the
one for version - which is constant locally, but seems rare in the CI.

It will likely be better to add this and the secret source to the
output, but this can be done in a separate PR with testing, once the CI
and tests are stable.